# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Validacion Valeria Barona",
        "description": "Creacion de usuarios - Documentación",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
        "/auth/do_register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Registro de usuario",
                "description": "Realiza el registro de un nuevo usuario.",
                "operationId": "bf608af1f2d97cc0b40e1d24c29c6088",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nombre del Usuario"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "usuario@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "contraseña123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario registrado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario registrado correctamente"
                                        },
                                        "user": {
                                            "type": "object",
                                            "example": {
                                                "id": 1,
                                                "name": "Nombre del Usuario",
                                                "email": "usuario@example.com",
                                                "created_at": "2023-02-23T00:09:16.000000Z",
                                                "updated_at": "2023-02-23T00:09:16.000000Z"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos de registro inválidos"
                    }
                }
            }
        },
        "/auth/do_login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Iniciar sesión",
                "description": "Realiza la autenticación del usuario.",
                "operationId": "d73c5a1b5df318508ee13d592e3666e2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user": {
                                        "type": "string",
                                        "example": "usuario@example.com"
                                    },
                                    "pwd": {
                                        "type": "string",
                                        "example": "contraseña123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inicio de sesión exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "ok"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "jwt_token_here"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Credenciales inválidas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "result": "error",
                                                "error_id": 422,
                                                "error_msg": "Algunos de los datos introducidos no son correctos"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/list": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Obtiene la lista de usuarios",
                "operationId": "b7e0f885a67ecac3dc98d53e3440bfd8",
                "responses": {
                    "200": {
                        "description": "Lista de usuarios recuperada con éxito"
                    }
                }
            }
        },
        "/users/store": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Crea un nuevo usuario",
                "description": "Almacenar un nuevo usuario",
                "operationId": "3a7352f89022edbe32db59b9d20d2fa6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nombre del Usuario"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "usuario@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "contraseña123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Nombre del Usuario"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "usuario@example.com"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "UNPROCESSABLE ENTITY",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/show/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Obtiene la información de un usuario",
                "operationId": "c48e1527c20eeadf088335a406c84d9e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario recuperado con éxito"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                }
            }
        },
        "/users/update/{id}": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Actualiza la información de un usuario",
                "operationId": "cbd1b94968e45761b4f8417868d20bd5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maximum": 255
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "minimum": 8
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario actualizado con éxito"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                }
            }
        },
        "/users/destroy/{id}": {
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Elimina un usuario",
                "operationId": "9cb4154c0e1476c3a6eefb712a9f0187",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario eliminado correctamente"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                }
            }
        }
    }
}